Move FEAT-specific mesh-cropping, adaptive kriging search strategy, and parameterization functions to `FEAT` subpackage.
