+++ echopop/nwfsc_feat/FEAT/__init__.py
@@ -1,0 +1,12 @@
+from .fun import (
+    get_survey_western_extents, 
+    western_boundary_search_strategy
+)
+from . import parameters
+
+__all__ = [
+    "get_survey_western_extents",
+    "parameters",
+    "transect_ends_crop",
+    "western_boundary_search_strategy",
+]
