+++ echopop/nwfsc_feat/kriging.py
@@ -98,167 +98,6 @@ def uniform_search_strategy(
     # Return Tuple
     return wr_indices, oos_indices, oos_weights
 
-
-def western_boundary_search_strategy(
-    kriging_mesh: pd.DataFrame,
-    western_extent: pd.DataFrame,
-    coordinate_names: Tuple[str, str],
-    sparse_radii: np.ndarray[int],
-    valid_distances: np.ndarray[int],
-    local_points: np.ndarray[float],
-    distance_matrix_masked: np.ndarray[float],
-    nearby_indices: np.ndarray[int],
-    k_min: int,
-    k_max: int,
-    search_radius: float,
-    wr_indices: np.ndarray[int],
-    oos_indices: np.ndarray[np.number],
-    oos_weights: np.ndarray[float],
-    **kwargs,
-) -> Tuple[np.ndarray[np.number], np.ndarray[np.number], np.ndarray[np.number]]:
-    """
-    Search strategy that applies western boundary constraints for transect-based surveys
-
-    Parameters
-    ----------
-    kriging_mesh : pd.DataFrame
-        Kriging mesh used for interpolated data values via geostatistics.
-    western_extent : pd.DataFrame
-        DataFrame with the western-most extent of each transect line used for re-weighting the
-        out-of-sample/extrapolated kriged values.
-    coordinate_names : Tuple[str, str]
-        Names of the coordinate columns when using DataFrames. Expected format: (x_col, y_col).
-    sparse_radii : np.ndarray[int]
-        Indices where there are fewer than `k_min` nearest neighbors.
-    valid_distances : np.ndarray[int]
-        The number of masked distance matrix values where extrapolation is required.
-    local_points : np.ndarray[float]
-        An array with the sorted distances (from nearest to furthest) relative to each point.
-    distance_matrix_masked : np.ndarray[float]
-        An array with the search-radius-masked nearest neighbor distances.
-    nearby_indices : np.ndarray[int]
-        Indices of points that require extrapolation.
-    k_min : int
-        The minimum number of nearest neighbors required for including values for kriging within
-        the search radius.
-    k_max : int
-        The maximum number of nearest neighbors required for including values for kriging detected
-        within the search radius.
-    search_radius : float
-        The adaptive search radius that identifies the *k*-nearest neighbors around each
-        georeferenced value that are subsequently kriged.
-    wr_indices : np.ndarray[int]
-        Indices of within-radius (WR) (i.e. < `k_max`) points.
-    oos_indices : np.ndarray[np.number]
-        Template array based on the size of the data input and `k_min` that will contain indices
-        where extrapolation is required where there are fewer than `k_min` nearest neighbors.
-    oos_weights : np.ndarray[float]
-        Weights applied to extraplolated values.
-
-    Returns
-    -------
-    Tuple[np.ndarray[np.number], np.ndarray[np.number], np.ndarray[np.number]]
-        A tuple with updated values for `wr_indices`, `oos_indices`, and `oos_weights` via a
-        search strategy that uses an extrapolation re-weighting based on transect extents.
-    """
-
-    # Parse ordered coordinate names
-    x_name, y_name = coordinate_names
-
-    # Index the mesh grid coordinates for bounding the search radius expansion/extrapolation
-    # ---- y-coordinates with array transformation to access matrix operations
-    mesh_y = kriging_mesh[y_name].to_numpy()[sparse_radii].reshape(-1, 1)
-    # ---- x-coordinates
-    mesh_x = kriging_mesh[x_name].to_numpy()[sparse_radii]
-
-    # Calculate the mesh distance from the western boundary of the survey transects
-    # ---- Find closest point
-    mesh_western_distance = np.abs(mesh_y - western_extent["y"].to_numpy()).argmin(axis=1)
-    # ---- Calculate the western limits (x-axis)
-    western_limit = western_extent.iloc[np.ravel(mesh_western_distance)][x_name]
-    # ---- Compute bounding threshold (for tapered extrapolation function)
-    western_threshold = western_limit - search_radius
-    # ---- Create a thresholded mask for lazy operations
-    western_limit_mask = mesh_x < western_threshold
-
-    # Adjust values that don't fall outside the western extent
-    if np.any(~western_limit_mask):
-        # ---- Grab all values that don't fall outside the western extent
-        soft_extrapolation_index = sparse_radii[~western_limit_mask]
-        # ---- Find the local points where there are at least some valid points
-        partial_indices = soft_extrapolation_index[valid_distances[soft_extrapolation_index] > 0]
-        # ---- Update the current values in `wr_indices`
-        if len(partial_indices) > 0:
-            # -------- Create boolean mask for within-range/sample points
-            wr_mask = local_points[partial_indices, :k_max] < search_radius
-            # -------- Create temporary matrix for within-range samples
-            wr_tmp = wr_indices[partial_indices].copy()
-            # -------- Create temporary matrix for oos samples
-            oos_tmp = wr_indices[partial_indices].copy()
-            # -------- Update temporary matrix by applying `wr_mask` for wr points
-            wr_tmp[~wr_mask] = np.nan
-            # -------- Update temporary matrix by applying `wr_mask` for oos points
-            oos_tmp[wr_mask] = np.nan
-            # -------- Assign the OOS values to `oos_indices`
-            oos_indices[partial_indices] = np.sort(oos_tmp[:, :k_min])
-            # -------- Apply the mask to the remaining `wr_indices` values
-            wr_indices[partial_indices] = np.sort(wr_tmp[:, :k_max])
-
-        # ---- Find the local points where there are no valid points within the search radius
-        full_extrap_indices = soft_extrapolation_index[
-            valid_distances[soft_extrapolation_index] == 0
-        ]
-        if len(full_extrap_indices) > 0:
-            # -------- Update `oos_indices`
-            oos_indices[full_extrap_indices] = wr_indices[full_extrap_indices, :k_min]
-            # -------- Update `wr_indices`
-            wr_indices[full_extrap_indices] = np.nan
-
-    # Taper function for extrapolating values outside the search radius
-    if np.any(western_limit_mask):
-        # ---- Index these values
-        extrapolation_index = sparse_radii[western_limit_mask]
-        # ---- Compute the OOS kriging weights
-        oos_mean = np.apply_along_axis(np.nanmean, 1, local_points[extrapolation_index, :k_min])
-        # ---- Exponentiate the OOS mean
-        oos_exp = np.exp(-oos_mean / search_radius)
-        # ---- Update the OOS weights
-        oos_weights[extrapolation_index] = oos_exp
-        # ---- Get the outside indices that correspond to this tapered extrapolation
-        sparse_extrapolation_index = nearby_indices[western_limit_mask].astype(float)
-        # ---- Apply indices as a mask to the NaN-masked distance matrix
-        extrapolated_distance = np.take_along_axis(
-            distance_matrix_masked[sparse_radii][western_limit_mask],
-            sparse_extrapolation_index.astype(int),
-            axis=1,
-        )
-        # ---- Create NaN mask
-        extrapolated_nan_mask = ~np.isnan(extrapolated_distance)
-        # -------- Apply mask to indices
-        sparse_extrapolation_index_nan = sparse_extrapolation_index.copy()
-        sparse_extrapolation_index_nan[extrapolated_nan_mask] = np.nan
-        # -------- Update `out_of_sample_indices` matrix
-        oos_indices[extrapolation_index] = np.sort(sparse_extrapolation_index_nan)
-        # ---- Get inside indices that apply to these points
-        # -------- Create NaN mask for within-sample values
-        interpolated_nan_mask = np.isnan(extrapolated_distance)
-        # -------- Apply mask to indices
-        sparse_interpolation_index_nan = sparse_extrapolation_index.copy()
-        sparse_interpolation_index_nan[interpolated_nan_mask] = np.nan
-        # -------- Pad NaN to match `within_sample_indices` matrix
-        sparse_interpolation_pad = np.pad(
-            sparse_interpolation_index_nan,
-            [(0, 0), (0, k_max - k_min)],
-            mode="constant",
-            constant_values=np.nan,
-        )
-        # -------- Updated `within_sample_indices` matrix
-        wr_indices[extrapolation_index] = np.sort(sparse_interpolation_pad)
-
-    # Return Tuple
-    return wr_indices, oos_indices, oos_weights
-
-
 def search_radius_mask(distance_matrix: np.ndarray, search_radius: float) -> np.ndarray[float]:
     """
     Generate a mask for the search radius to identify points within the search distance.
